requirements.txt
streamlit
pandas
flask
flask-sqlalchemy
psycopg2-binary
